benin_luxe_cajou_api/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-Marshmallow
marshmallow-sqlalchemy
PyMySQL
python-dotenv
bcrypt
cloudinary
gunicorn
Flask-Mail